chore: restructure ACIR-gen code
